backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
beautifulsoup4==4.12.2
requests==2.31.0
python-whois==0.8.0
gunicorn==22.0.0
